Found multiple main methods, moved one of them out into another bin for the crate. Convert sys arg to path.
